Refactor server loader v3 admin api controller.
